Update toml-test
This updates the test files from upstream; all additional tests are now
also integrated upstream, so I removed them here. The only exception is
the one that subseconds are truncated to microseconds as that's
implementation defined behaviour, so I added a new unit test here.

The repo was moved from BurntSushi to toml-lang, so renamed a few things
to reflect that. The iarna and "compliance" tests are no longer used, so
simplify the code a bit as that doesn't need to be supported.

Make the script executable so it can be run with the toml-test runner:

	% PYTHONPATH=./src toml-test ./tests/tomltest.py
	toml-test v2023-10-13 [./tests/tomltest.py]: using embedded tests
	  valid tests: 166 passed,  0 failed
	invalid tests: 355 passed,  0 failed

This gives slightly nicer errors (IMHO), and also selfishly makes it a
bit easier to keep the [test matrix] updated, as I won't have to include
my own script and can just run this one.

[test matrix]: https://arp242.github.io/toml-test-matrix